Move ChangeMapObjectsTile to changemapobject files
